16/16.test.ts
import * as path from "@std/path";
import { expect } from "@std/expect";
import { describe, it } from "@std/testing/bdd";
import { solveA, solveB } from "./16.ts";

const currentDay = path.basename(Deno.cwd());

describe(`AOC 2024 - Day ${currentDay}`, () => {
	describe("Part A", () => {
		it("Example", () => {
			expect(solveA("example_a", currentDay)).toBe(11048);
		});

		it("Solution", () => {
			expect(solveA("input", currentDay)).toBe(105508);
		});
	});

	describe("Part B", () => {
		it("Example", () => {
			expect(solveB("example_b", currentDay)).toBe(64);
		});

		it("Solution", () => {
			expect(solveB("input", currentDay)).toBe(548);
		});
	});
});
